pyproject.toml
[tool.poetry]
name = "ws"
version = "0.1.0"
description = "The simplest websocket cli"
authors = ["le_woudar <lewoudar@gmail.com>"]
readme = "README.md"
license = "Apache-2.0"
keywords = ["websocket", "cli", "trio"]
repository = "https://github.com/lewoudar/certipie"

classifiers = [
    "Intended Audience :: Developers",
    "Topic :: Terminals",
    "Operating System :: OS Independent"
]

packages = [
    { include = "ws" }
]

[tool.poetry.dependencies]
python = "^3.7"
trio-websocket = "^0.9.2"
anyio = "^3.5.0"
rich = "^12.0.0"
pydantic = "^1.9.0"
tomli = "^2.0.1"
click = "^8.0.4"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"
pytest-cov = "^3.0.0"
pytest-trio = "^0.7.0"
pre-commit = "^2.17.0"
isort = "^5.10.1"
bandit = "^1.7.4"
flake8 = "^4.0.1"
cffi = "^1.15.0"
black = "^22.1.0"
pytest-mock = "^3.7.0"
mock = "^4.0.3"

[tool.poetry.scripts]
ws = "ws.main:cli"

[tool.pytest.ini_options]
testpaths = ["tests"]
# addopts = "--cov=ws --cov-report html --cov-report xml"
trio_mode = true

[tool.isort]
line_length = 120
profile = "black"

[tool.black]
line-length = 120
skip-string-normalization = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
